Refactored locations of functions and added log_driver_version
